+++ apps/files_sharing/tests/sharedmount.php
@@ -1,31 +1,38 @@
 <?php
 /**
- * ownCloud
+ * @author Björn Schießle <schiessle@owncloud.com>
+ * @author Joas Schilling <nickvergessen@owncloud.com>
+ * @author Morris Jobke <hey@morrisjobke.de>
+ * @author Robin Appelman <icewind@owncloud.com>
+ * @author Roeland Jago Douma <rullzer@owncloud.com>
+ * @author Thomas Müller <thomas.mueller@tmit.eu>
+ * @author Vincent Petry <pvince81@owncloud.com>
  *
- * @author Bjoern Schiessle
- * @copyright 2014 Bjoern Schiessle <schiessle@owncloud.com>
+ * @copyright Copyright (c) 2016, ownCloud, Inc.
+ * @license AGPL-3.0
  *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU AFFERO GENERAL PUBLIC LICENSE
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or any later version.
+ * This code is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
  *
- * This library is distributed in the hope that it will be useful,
+ * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU AFFERO GENERAL PUBLIC LICENSE for more details.
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
  *
- * You should have received a copy of the GNU Affero General Public
- * License along with this library.  If not, see <http://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU Affero General Public License, version 3,
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>
  *
  */
 
 /**
  * Class Test_Files_Sharing_Api
+ *
+ * @group DB
  */
 class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 
-	function setUp() {
+	protected function setUp() {
 		parent::setUp();
 
 		$this->folder = '/folder_share_storage_test';
@@ -40,9 +47,15 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 		$this->view->file_put_contents($this->folder . $this->filename, "file in subfolder");
 	}
 
-	function tearDown() {
-		$this->view->unlink($this->folder);
-		$this->view->unlink($this->filename);
+	protected function tearDown() {
+		if ($this->view) {
+			if ($this->view->file_exists($this->folder)) {
+				$this->view->unlink($this->folder);
+			}
+			if ($this->view->file_exists($this->filename)) {
+				$this->view->unlink($this->filename);
+			}
+		}
 
 		parent::tearDown();
 	}
@@ -50,59 +63,46 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 	/**
 	 * test if the mount point moves up if the parent folder no longer exists
 	 */
-	function testShareMountLoseParentFolder() {
+	public function testShareMountLoseParentFolder() {
 
 		// share to user
-		$fileinfo = $this->view->getFileInfo($this->folder);
-		$result = \OCP\Share::shareItem('folder', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER,
-			self::TEST_FILES_SHARING_API_USER2, 31);
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_USER,
+			$this->folder,
+			self::TEST_FILES_SHARING_API_USER1,
+			self::TEST_FILES_SHARING_API_USER2,
+			\OCP\Constants::PERMISSION_ALL);
 
-		$statement = "UPDATE `*PREFIX*share` SET `file_target` = ? where `share_with` = ?";
-		$query = \OC_DB::prepare($statement);
-		$arguments = array('/foo/bar' . $this->folder, self::TEST_FILES_SHARING_API_USER2);
-		$query->execute($arguments);
+		$share->setTarget('/foo/bar' . $this->folder);
+		$this->shareManager->moveShare($share, self::TEST_FILES_SHARING_API_USER2);
 
-		$query = \OC_DB::prepare('SELECT * FROM `*PREFIX*share`');
-		$result = $query->execute();
-
-		$shares = $result->fetchAll();
-
-		$this->assertSame(1, count($shares));
-
-		$share = reset($shares);
-		$this->assertSame('/foo/bar' . $this->folder, $share['file_target']);
+		$share = $this->shareManager->getShareById($share->getFullId());
+		$this->assertSame('/foo/bar' . $this->folder, $share->getTarget());
 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
-
 		// share should have moved up
 
-		$query = \OC_DB::prepare('SELECT * FROM `*PREFIX*share`');
-		$result = $query->execute();
-
-		$shares = $result->fetchAll();
-
-		$this->assertSame(1, count($shares));
-
-		$share = reset($shares);
-		$this->assertSame($this->folder, $share['file_target']);
+		$share = $this->shareManager->getShareById($share->getFullId());
+		$this->assertSame($this->folder, $share->getTarget());
 
 		//cleanup
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
-		\OCP\Share::unshare('folder', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER, self::TEST_FILES_SHARING_API_USER2);
+		$this->shareManager->deleteShare($share);
 		$this->view->unlink($this->folder);
 	}
 
 	/**
 	 * @medium
 	 */
-	function testDeleteParentOfMountPoint() {
-
+	public function testDeleteParentOfMountPoint() {
 		// share to user
-		$fileinfo = $this->view->getFileInfo($this->folder);
-		$result = \OCP\Share::shareItem('folder', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER,
-			self::TEST_FILES_SHARING_API_USER2, 31);
-
-		$this->assertTrue($result);
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_USER,
+			$this->folder,
+			self::TEST_FILES_SHARING_API_USER1,
+			self::TEST_FILES_SHARING_API_USER2,
+			\OCP\Constants::PERMISSION_ALL
+		);
 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
 		$user2View = new \OC\Files\View('/' . self::TEST_FILES_SHARING_API_USER2 . '/files');
@@ -134,39 +134,54 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 		$this->view->unlink($this->folder);
 	}
 
-	function testMoveSharedFile() {
-		$fileinfo = $this->view->getFileInfo($this->filename);
-		$result = \OCP\Share::shareItem('file', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER,
-			self::TEST_FILES_SHARING_API_USER2, 31);
+	public function testMoveSharedFile() {
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_USER,
+			$this->filename,
+			self::TEST_FILES_SHARING_API_USER1,
+			self::TEST_FILES_SHARING_API_USER2,
+			\OCP\Constants::PERMISSION_READ | \OCP\Constants::PERMISSION_UPDATE | \OCP\Constants::PERMISSION_SHARE
+		);
 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
 
-		\OC\Files\Filesystem::rename($this->filename, "newFileName");
+		\OC\Files\Filesystem::rename($this->filename, $this->filename . '_renamed');
 
-		$this->assertTrue(\OC\Files\Filesystem::file_exists('newFileName'));
+		$this->assertTrue(\OC\Files\Filesystem::file_exists($this->filename . '_renamed'));
 		$this->assertFalse(\OC\Files\Filesystem::file_exists($this->filename));
 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
 		$this->assertTrue(\OC\Files\Filesystem::file_exists($this->filename));
-		$this->assertFalse(\OC\Files\Filesystem::file_exists("newFileName"));
+		$this->assertFalse(\OC\Files\Filesystem::file_exists($this->filename . '_renamed'));
+
+		// rename back to original name
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
+		\OC\Files\Filesystem::rename($this->filename . '_renamed', $this->filename);
+		$this->assertFalse(\OC\Files\Filesystem::file_exists($this->filename . '_renamed'));
+		$this->assertTrue(\OC\Files\Filesystem::file_exists($this->filename));
 
 		//cleanup
-		\OCP\Share::unshare('file', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER, self::TEST_FILES_SHARING_API_USER2);
+		$this->shareManager->deleteShare($share);
 	}
 
 	/**
 	 * share file with a group if a user renames the file the filename should not change
 	 * for the other users
 	 */
-	function testMoveGroupShare () {
+	public function testMoveGroupShare () {
 		\OC_Group::createGroup('testGroup');
 		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER1, 'testGroup');
 		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER2, 'testGroup');
 		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
 
 		$fileinfo = $this->view->getFileInfo($this->filename);
-		$result = \OCP\Share::shareItem('file', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_GROUP,
-			"testGroup", 31);
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_GROUP,
+			$this->filename,
+			self::TEST_FILES_SHARING_API_USER1,
+			'testGroup',
+			\OCP\Constants::PERMISSION_READ | \OCP\Constants::PERMISSION_UPDATE | \OCP\Constants::PERMISSION_SHARE
+		);
 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
 
@@ -186,7 +201,8 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 		$this->assertFalse(\OC\Files\Filesystem::file_exists("newFileName"));
 
 		//cleanup
-		\OCP\Share::unshare('file', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_GROUP, 'testGroup');
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
+		$this->shareManager->deleteShare($share);
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER1, 'testGroup');
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER2, 'testGroup');
 		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
@@ -198,7 +214,7 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 	 * @param string $expectedResult
 	 * @param bool $exception if a exception is expected
 	 */
-	function testStripUserFilesPath($path, $expectedResult, $exception) {
+	public function testStripUserFilesPath($path, $expectedResult, $exception) {
 		$testClass = new DummyTestClassSharedMount(null, null);
 		try {
 			$result = $testClass->stripUserFilesPathDummy($path);
@@ -212,7 +228,7 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 		}
 	}
 
-	function dataProviderTestStripUserFilesPath() {
+	public function dataProviderTestStripUserFilesPath() {
 		return array(
 			array('/user/files/foo.txt', '/foo.txt', false),
 			array('/user/files/folder/foo.txt', '/folder/foo.txt', false),
@@ -221,6 +237,188 @@ class Test_Files_Sharing_Mount extends OCA\Files_sharing\Tests\TestCase {
 			array('/files/foo.txt', null, true),
 			array('/foo.txt', null, true),
 		);
+	}
+
+	public function dataPermissionMovedGroupShare() {
+		$data = [];
+
+		$powerset = function($permissions) {
+			$results = [\OCP\Constants::PERMISSION_READ];
+
+			foreach ($permissions as $permission) {
+				foreach ($results as $combination) {
+					$results[] = $permission | $combination;
+				}
+			}
+			return $results;
+		};
+
+		//Generate file permissions
+		$permissions = [
+			\OCP\Constants::PERMISSION_UPDATE,
+			\OCP\Constants::PERMISSION_SHARE,
+		];
+
+		$allPermissions = $powerset($permissions);
+
+		foreach ($allPermissions as $before) {
+			foreach ($allPermissions as $after) {
+				if ($before === $after) { continue; }
+
+				$data[] = [
+					'file', 
+					$before,
+					$after,
+				];
+			}
+		}
+
+		//Generate folder permissions
+		$permissions = [
+			\OCP\Constants::PERMISSION_UPDATE,
+			\OCP\Constants::PERMISSION_CREATE,
+			\OCP\Constants::PERMISSION_SHARE,
+			\OCP\Constants::PERMISSION_DELETE,
+		];
+
+		$allPermissions = $powerset($permissions);
+
+		foreach ($allPermissions as $before) {
+			foreach ($allPermissions as $after) {
+				if ($before === $after) { continue; }
+
+				$data[] = [
+					'folder',
+					$before,
+					$after,
+				];
+			}
+		}
+
+		return $data;
+	}
+
+
+
+	/**
+	 * moved mountpoints of a group share should keep the same permission as their parent group share.
+	 * See #15253
+	 *
+	 * @dataProvider dataPermissionMovedGroupShare
+	 */
+	function testPermissionMovedGroupShare($type, $beforePerm, $afterPerm) {
+
+		if ($type === 'file') {
+			$path = $this->filename;
+		} else if ($type === 'folder') {
+			$path = $this->folder;
+		}
+
+		\OC_Group::createGroup('testGroup');
+		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER1, 'testGroup');
+		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER2, 'testGroup');
+		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
+
+		// Share item with group
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_GROUP,
+			$path,
+			self::TEST_FILES_SHARING_API_USER1,
+			'testGroup',
+			$beforePerm
+		);
+
+		// Login as user 2 and verify the item exists
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
+		$this->assertTrue(\OC\Files\Filesystem::file_exists($path));
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER2);
+		$this->assertEquals($beforePerm, $result->getPermissions());
+
+		// Now move the item forcing a new entry in the share table
+		\OC\Files\Filesystem::rename($path, "newPath");
+		$this->assertTrue(\OC\Files\Filesystem::file_exists('newPath'));
+		$this->assertFalse(\OC\Files\Filesystem::file_exists($path));
+
+		// change permissions
+		$share->setPermissions($afterPerm);
+		$this->shareManager->updateShare($share);
+
+		// Login as user 3 and verify that the permissions are changed
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER3);
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER3);
+		$this->assertNotEmpty($result);
+		$this->assertEquals($afterPerm, $result->getPermissions());
+
+		// Login as user 2 and verify that the permissions are changed
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER2);
+		$this->assertNotEmpty($result);
+		$this->assertEquals($afterPerm, $result->getPermissions());
+		$this->assertEquals('/newPath', $result->getTarget());
+
+		//cleanup
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
+		$this->shareManager->deleteShare($share);
+		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER1, 'testGroup');
+		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER2, 'testGroup');
+		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
+	}
+
+	/**
+	 * If the permissions on a group share are upgraded be sure to still respect 
+	 * removed shares by a member of that group
+	 */
+	function testPermissionUpgradeOnUserDeletedGroupShare() {
+		\OC_Group::createGroup('testGroup');
+		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER1, 'testGroup');
+		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER2, 'testGroup');
+		\OC_Group::addToGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
+
+		$connection = \OC::$server->getDatabaseConnection();
+
+		// Share item with group
+		$fileinfo = $this->view->getFileInfo($this->folder);
+		$share = $this->share(
+			\OCP\Share::SHARE_TYPE_GROUP,
+			$this->folder,
+			self::TEST_FILES_SHARING_API_USER1,
+			'testGroup',
+			\OCP\Constants::PERMISSION_READ
+		);
+
+		// Login as user 2 and verify the item exists
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
+		$this->assertTrue(\OC\Files\Filesystem::file_exists($this->folder));
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER2);
+		$this->assertNotEmpty($result);
+		$this->assertEquals(\OCP\Constants::PERMISSION_READ, $result->getPermissions());
+
+		// Delete the share
+		$this->assertTrue(\OC\Files\Filesystem::rmdir($this->folder));
+		$this->assertFalse(\OC\Files\Filesystem::file_exists($this->folder));
+
+		// Verify we do not get a share
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER2);
+		$this->assertEquals(0, $result->getPermissions());
+
+		// Login as user 1 again and change permissions
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
+		$share->setPermissions(\OCP\Constants::PERMISSION_ALL);
+		$share = $this->shareManager->updateShare($share);
+
+		// Login as user 2 and verify 
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
+		$this->assertFalse(\OC\Files\Filesystem::file_exists($this->folder));
+		$result = $this->shareManager->getShareById($share->getFullId(), self::TEST_FILES_SHARING_API_USER2);
+		$this->assertEquals(0, $result->getPermissions());
+
+		$this->shareManager->deleteShare($share);
+
+		//cleanup
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
+		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER1, 'testGroup');
+		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER2, 'testGroup');
+		\OC_Group::removeFromGroup(self::TEST_FILES_SHARING_API_USER3, 'testGroup');
 	}
 
 }
